+++ src/main/kotlin/racingcar/RacingApplication.kt
@@ -2,8 +2,8 @@ package racingcar
 
 import racingcar.tasks.WinnersReader
 import racingcar.tasks.Game
-import racingcar.ui.InputView
-import racingcar.ui.OutputView
+import racingcar.view.InputView
+import racingcar.view.OutputView
 
 fun main() {
     run(Game())
