+++ src/test/kotlin/racingcar/tasks/GameTest.kt
@@ -2,7 +2,7 @@ package racingcar.tasks
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import racingcar.model.Car
+import racingcar.domain.Car
 
 internal class GameTest {
